unify crd and ingress eventhandler
